añadi un autocorregidor de contraseñas hash

+++ src/php/deleteUser.php
@@ -25,7 +25,15 @@ if($conex) {
         while ($li = $result->fetch_assoc()) {
             $hash = $li['Contra'];
         }
-        if (password_verify($pass, $hash)){
+        function esHash($str) {
+            return preg_match('/^\$2[aby]\$\d{2}\$[\.\/A-Za-z0-9]{53}$/', $str);
+        }
+        $is_hash = esHash($hash);
+        if (!$is_hash && $pass == $hash) {
+                $hash = password_hash($pass, PASSWORD_DEFAULT);
+                $is_hash = true;
+        }
+        if ($is_hash && password_verify($pass, $hash)){
             $stmt = $conex->prepare("DELETE FROM usuario WHERE Nomb = ? AND Edad = ?");
             $stmt->bind_param("si", $nombre, $edad);
             $stmt->execute();
